# Vectorize IndexOf for OrdinalIgnoreCase