.. when the value starts with any ASCII symbol (even if then it contains non-ASCII too). This implementation is based on @stephentoub's trick used in Regex

Benchmark:
```csharp
StringComparison CmpMode = StringComparison.OrdinalIgnoreCase;

////////////////////////////////////////////////////////////////////////////////

string SmallStr = "hello";

[Benchmark]
public int IndexOf_l() => SmallStr.IndexOf("l", CmpMode);

////////////////////////////////////////////////////////////////////////////////

string DotnetHeader =
    ".NET Runtime uses third-party libraries or other resources that may be distributed";

[Benchmark]
public int IndexOf_distr() => DotnetHeader.IndexOf("distr", CmpMode);

////////////////////////////////////////////////////////////////////////////////

string ThirdPartyNotice = File.ReadAllText(@"C:\prj\runtime\THIRD-PARTY-NOTICES.TXT");

[Benchmark]
public int IndexOf_JavaScript() => ThirdPartyNotice.IndexOf("JavaScript", CmpMode);

[Benchmark]
public int IndexOf_Percentage() => ThirdPartyNotice.AsSpan().IndexOf("%", CmpMode);
```

|             Method |                   Toolchain |           Mean |       Error |      StdDev | Ratio |
|------------------- |---------------------------- |---------------:|------------:|------------:|------:|
|          IndexOf_l |      \Core_Root\corerun.exe |     10.942 ns |   0.0905 ns |   0.0847 ns |  1.00 |
|          IndexOf_l | \Core_Root_base\corerun.exe |      12.751 ns |   0.1634 ns |   0.1448 ns |  **1.17** |
|                    |                             |                |             |             |       |
|      IndexOf_distr |      \Core_Root\corerun.exe |      17.964 ns |   0.0624 ns |   0.0521 ns |  1.00 |
|      IndexOf_distr | \Core_Root_base\corerun.exe |     169.188 ns |   1.0764 ns |   1.0069 ns |  **9.43** |
|                    |                             |                |             |             |       |
| IndexOf_JavaScript |      \Core_Root\corerun.exe |   2,635.342 ns |  21.6150 ns |  19.1611 ns |  1.00 |
| IndexOf_JavaScript | \Core_Root_base\corerun.exe |  97,961.531 ns | 359.4873 ns | 280.6641 ns | **37.23** |
|                    |                             |                |             |             |       |
| IndexOf_Percentage |      \Core_Root\corerun.exe |   1,648.208 ns |  12.4174 ns |  11.6152 ns |  1.00 |
| IndexOf_Percentage | \Core_Root_base\corerun.exe | 113,032.308 ns | 289.2747 ns | 225.8467 ns | **68.57** |