@danmoseley 2022-04-08
It looks like our perf test coverage of this scenario is exactly one case, which is finding a 4 letter ASCII string in a ~200 char ASCII string where it is not present. 
https://github.com/dotnet/performance/blob/a37c8e3e4c72d584934e97a1fdc0a4785ddb8674/src/benchmarks/micro/libraries/System.Globalization/StringSearch.cs#L100

It might be worth adding a couple more so we'd notice if this change regressed.