@DrewScoggins 2022-04-14
Win-Arm64 Improvement: https://github.com/dotnet/perf-autofiling-issues/issues/4622