@stephentoub 2022-04-14
That's a pretty graph.