@dakersnar 2022-04-28
Windows x86 Improvements: https://github.com/dotnet/perf-autofiling-issues/issues/4602