@dakersnar 2022-05-13
Huge improvements on all configs!

<details>

## System.Globalization.Tests.StringSearch.IndexOf_Word_NotFound(Options: (en-US, OrdinalIgnoreCase, False))

| Result |Ratio | Operating System      | Bit   | Processor Name                                  |
| ------ |-----:| --------------------- | ----- | ----------------------------------------------- |
| Faster | 8.89 | ubuntu 20.04          | Arm64 | Unknown processor                               |
| Faster |10.06 | Windows 10            | Arm64 | Microsoft SQ1 3.0 GHz                           |
| Faster | 9.66 | Windows 11            | Arm64 | Microsoft SQ1 3.0 GHz                           |
| Faster | 9.91 | Windows 11            | Arm64 | Unknown processor                               |
| Faster |11.87 | macOS Monterey 12.3   | Arm64 | Apple M1 Max                                    |
| Faster |20.18 | Windows 10            | X64   | Intel Xeon Platinum 8272CL CPU 2.60GHz          |
| Faster |12.90 | Windows 10            | X64   | Intel Xeon CPU E5-1650 v4 3.60GHz               |
| Faster |14.94 | Windows 10            | X64   | Intel Core i7-6700 CPU 3.40GHz (Skylake)        |
| Faster |12.63 | Windows 11            | X64   | AMD Ryzen Threadripper PRO 3945WX 12-Cores      |
| Faster |13.03 | Windows 11            | X64   | AMD Ryzen 9 5950X                               |
| Faster |20.77 | Windows 11            | X64   | Intel Core i7-8700 CPU 3.20GHz (Coffee Lake)    |
| Faster |18.22 | Windows 11            | X64   | Intel Core i9-9900T CPU 2.10GHz                 |
| Faster |14.99 | alpine 3.13           | X64   | Intel Core i7-7700 CPU 3.60GHz (Kaby Lake)      |
| Faster |19.26 | pop 22.04             | X64   | Intel Core i7-6600U CPU 2.60GHz (Skylake)       |
| Faster |16.62 | ubuntu 18.04          | X64   | Intel Xeon CPU E5-1650 v4 3.60GHz               |
| Faster |12.60 | ubuntu 18.04          | X64   | Intel Core i7-2720QM CPU 2.20GHz (Sandy Bridge) |
| Faster |14.98 | ubuntu 20.04          | X64   | AMD Ryzen 9 5900X                               |
| Faster |17.32 | ubuntu 20.04          | X64   | Intel Core i7-8700 CPU 3.20GHz (Coffee Lake)    |
| Faster |12.09 | Windows 10            | X86   | Intel Xeon CPU E5-1650 v4 3.60GHz               |
| Faster |14.52 | Windows 10            | X86   | Intel Core i7-6700 CPU 3.40GHz (Skylake)        |
| Faster |15.70 | Windows 11            | X86   | AMD Ryzen Threadripper PRO 3945WX 12-Cores      |
| Faster |16.66 | macOS Monterey 12.2.1 | X64   | Intel Core i7-5557U CPU 3.10GHz (Broadwell)     |
| Faster |14.80 | macOS Monterey 12.3.1 | X64   | Intel Core i7-4870HQ CPU 2.50GHz (Haswell)      |

</details>